move functions to service
